# internal/platform/docs/swagger.yaml
basePath: /api
definitions:
  httptransport.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 小智服务端，包含OTA与Vision等接口
  title: 小智服务端 API 文档
  version: "1.0"
paths:
...
  /startup/workflows/{id}:
    get:
      description: 根据ID获取特定工作流的详细信息
      parameters:
      - description: 工作流ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 获取工作流详情
      tags:
      - Workflow
...
swagger: "2.0"